+++ src/ifs/wl_seat/zwp_virtual_keyboard_v1.rs
@@ -30,7 +30,7 @@ pub struct ZwpVirtualKeyboardV1 {
 impl ZwpVirtualKeyboardV1 {
     fn for_each_kb<F>(&self, mut f: F)
     where
-        F: FnMut(u32, &WlSurface, &WlKeyboard),
+        F: FnMut(u64, &WlSurface, &WlKeyboard),
     {
         let Some(surface) = self.seat.keyboard_node.get().node_into_surface() else {
             return;
