+++ src/tree/float.rs
@@ -632,7 +632,7 @@ impl Node for FloatNode {
         time_usec: u64,
         button: u32,
         state: KeyState,
-        _serial: u32,
+        _serial: u64,
     ) {
         if button != BTN_LEFT {
             return;
